// src/interface/pokemon.interface.tsx
export interface SimplePokemon {
    id     : string;
    name   : string;
    picture: string;
    color? : string;
}